Conflicts:
	src/generator_convolution_common.c
	src/generator_convolution_weight_update_avx512.c
	src/generator_x86_instructions.c
	src/libxsmm_dnn_setup.c

+++ src/generator_spgemm.c
@@ -280,12 +280,12 @@ void libxsmm_generator_spgemm( const char*                    i_file_out,
 
 #if !defined(NDEBUG)
     {
-      double *const l_tmp = (double*)malloc(l_row_count * l_column_count * sizeof(double));
+      double *const l_tmp = (double*)malloc((size_t)l_row_count * l_column_count * sizeof(double));
       unsigned int l_n;
       unsigned int l_m;
 
       /* mute static analysis about garbage content */
-      memset(l_tmp, 0, l_row_count * l_column_count * sizeof(double));
+      memset(l_tmp, 0, (size_t)l_row_count * l_column_count * sizeof(double));
 
       printf("CSC matrix data structure we just read:\n");
       printf("rows: %u, columns: %u, elements: %u\n", l_row_count, l_column_count, l_element_count);
@@ -346,12 +346,12 @@ void libxsmm_generator_spgemm( const char*                    i_file_out,
 
 #if !defined(NDEBUG)
     {
-      double *const l_tmp = (double*)malloc(l_row_count * l_column_count * sizeof(double));
+      double *const l_tmp = (double*)malloc((size_t)l_row_count * l_column_count * sizeof(double));
       unsigned int l_n;
       unsigned int l_m;
 
       /* mute static analysis about garbage content */
-      memset(l_tmp, 0, l_row_count * l_column_count * sizeof(double));
+      memset(l_tmp, 0, (size_t)l_row_count * l_column_count * sizeof(double));
 
       printf("CSR matrix data structure we just read:\n");
       printf("rows: %u, columns: %u, elements: %u\n", l_row_count, l_column_count, l_element_count);
